requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
Django==4.2.1
django-celery-beat==2.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
